moving to a proper python project setup
This is necessitating changing the imports to be correct
so that pytest can find things.
