+++ tests/test.py
@@ -1,9 +1,9 @@
 # This unitest test uses the following encoding: utf-8
 
-import extralifedonations
-import extralife_IO
-import donor
-import donation
+from eldonationtracker import extralifedonations
+from eldonationtracker import extralife_IO
+from eldonationtracker import donor
+from eldonationtracker import donation
 
 # run with command: py.test-3 extralifeunittests.py
 # or if in venv, just pytest extralifeunittests.py (works better)
